refactor: AmbientLight component
