+++ src/components/AmbientLight/AmbientLight.jsx
@@ -1,0 +1,11 @@
+import { useAmbientLightLogic } from './logic'
+
+function AmbientLight() {
+  const { ambient, intensityColor, colorAmbient } = useAmbientLightLogic()
+
+  return (
+    ambient && <ambientLight intensity={intensityColor} color={colorAmbient} />
+  )
+}
+
+export default AmbientLight
